requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
neo4j-driver==1.7.5
neobolt==1.7.15
neotime==1.7.4
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
